backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
openai
firebase_admin
python-multipart
aiofiles
cryptography
jwt
motor
firebase_admin